+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - White - [#63](https://github.com/dartoos-dev/eo_color/issues/63).
+- Accent color swatches - [#61](https://github.com/dartoos-dev/eo_color/issues/61).
 
 ## [0.0.11] - 2021-05-24
 ### Added